requirements.txt
django==1.2.3
South==0.7.2
django-annoying==0.7.4
django-debug-toolbar==0.8.3
wsgiref==0.1.2

